Rename controller and reconciler to conform to upstream.
