+++ pkg/reconciler/ingress/ingress.go
@@ -2,93 +2,99 @@ package ingress
 
 import (
 	"context"
-	"kourier/pkg/config"
 	"kourier/pkg/envoy"
+	"kourier/pkg/knative"
 
-	v1 "k8s.io/api/core/v1"
-
-	"k8s.io/apimachinery/pkg/types"
-
-	"knative.dev/serving/pkg/apis/networking"
-	"knative.dev/serving/pkg/reconciler"
-
-	"k8s.io/client-go/tools/cache"
-
-	kubeclient "knative.dev/pkg/client/injection/kube/client"
-	endpointsinformer "knative.dev/pkg/client/injection/kube/informers/core/v1/endpoints"
-	"knative.dev/pkg/configmap"
-	"knative.dev/pkg/controller"
-	"knative.dev/pkg/logging"
 	"knative.dev/pkg/tracker"
 
-	knativeclient "knative.dev/serving/pkg/client/injection/client"
-	ingressinformer "knative.dev/serving/pkg/client/injection/informers/networking/v1alpha1/ingress"
+	"knative.dev/serving/pkg/apis/networking/v1alpha1"
+
+	"knative.dev/pkg/network"
+
+	apierrors "k8s.io/apimachinery/pkg/api/errors"
+	"k8s.io/apimachinery/pkg/labels"
+	kubeclient "k8s.io/client-go/kubernetes"
+	corev1listers "k8s.io/client-go/listers/core/v1"
+	"k8s.io/client-go/tools/cache"
+	nv1alpha1lister "knative.dev/serving/pkg/client/listers/networking/v1alpha1"
 )
 
+// Values for the keys received in the reconciler. When they're not a standard
+// "namespace/name".
 const (
-	controllerName = "KourierController"
-	nodeID         = "3scale-kourier-gateway"
-	gatewayPort    = 19001
-	managementPort = 18000
+	FullResync = "full_resync"
 )
 
-func NewController(ctx context.Context, cmw configmap.Watcher) *controller.Impl {
-	kubernetesClient := kubeclient.Get(ctx)
-	knativeClient := knativeclient.Get(ctx)
+type Reconciler struct {
+	IngressLister   nv1alpha1lister.IngressLister
+	EndpointsLister corev1listers.EndpointsLister
+	EnvoyXDSServer  envoy.EnvoyXdsServer
+	kubeClient      kubeclient.Interface
+	CurrentCaches   *envoy.Caches
+	tracker         tracker.Interface
+}
 
-	envoyXdsServer := envoy.NewEnvoyXdsServer(
-		gatewayPort,
-		managementPort,
-		kubernetesClient,
-		knativeClient,
+func (reconciler *Reconciler) Reconcile(ctx context.Context, key string) error {
+	if key == FullResync {
+		return reconciler.fullReconcile()
+	}
+
+	namespace, name, err := cache.SplitMetaNamespaceKey(key)
+	if err != nil {
+		return err
+	}
+
+	ingress, err := reconciler.IngressLister.Ingresses(namespace).Get(name)
+	if apierrors.IsNotFound(err) {
+		reconciler.deleteIngress(namespace, name)
+		return nil
+	} else if err != nil {
+		return err
+	}
+
+	reconciler.updateIngress(ingress)
+	return nil
+}
+
+func (reconciler *Reconciler) fullReconcile() error {
+	ingresses, err := reconciler.IngressLister.List(labels.Everything())
+	if err != nil {
+		return err
+	}
+
+	kourierIngresses := knative.FilterByIngressClass(ingresses)
+
+	caches := reconciler.EnvoyXDSServer.SetSnapshotForIngresses(
+		nodeID,
+		kourierIngresses,
+		reconciler.EndpointsLister,
+		reconciler.tracker,
 	)
-	go envoyXdsServer.RunManagementServer()
-	go envoyXdsServer.RunGateway()
 
-	logger := logging.FromContext(ctx)
+	reconciler.CurrentCaches = caches
 
-	ingressInformer := ingressinformer.Get(ctx)
-	endpointsInformer := endpointsinformer.Get(ctx)
+	return nil
+}
 
-	caches := envoy.NewCaches()
+func (reconciler *Reconciler) deleteIngress(namespace, name string) {
+	reconciler.CurrentCaches.DeleteIngressInfo(name, namespace, reconciler.kubeClient)
+	reconciler.EnvoyXDSServer.SetSnapshotForCaches(reconciler.CurrentCaches, nodeID)
+}
 
-	c := &Reconciler{
-		IngressLister:   ingressInformer.Lister(),
-		EndpointsLister: endpointsInformer.Lister(),
-		EnvoyXDSServer:  envoyXdsServer,
-		kubeClient:      kubernetesClient,
-		CurrentCaches:   &caches,
-	}
+func (reconciler *Reconciler) updateIngress(ingress *v1alpha1.Ingress) {
+	envoy.UpdateInfoForIngress(
+		reconciler.CurrentCaches,
+		ingress,
+		reconciler.kubeClient,
+		reconciler.EndpointsLister,
+		network.GetClusterDomainName(),
+		reconciler.tracker,
+	)
 
-	impl := controller.NewImpl(c, logger, controllerName)
-	c.tracker = tracker.New(impl.EnqueueKey, controller.GetTrackerLease(ctx))
+	reconciler.EnvoyXDSServer.SetSnapshotForCaches(reconciler.CurrentCaches, nodeID)
 
-	// Force a first event to make sure we initialize a config. Otherwise, there
-	// will be no config until a Knative service is deployed.
-	// This is important because the gateway pods will not be marked as healthy
-	// until they have been able to fetch a config.
-	event := types.NamespacedName{
-		Name: FullResync,
-	}
-	impl.EnqueueKey(event)
-
-	// Ingresses need to be filtered by ingress class, so Kourier does not
-	// react to nor modify ingresses created by other gateways.
-	ingressInformerHandler := cache.FilteringResourceEventHandler{
-		FilterFunc: reconciler.AnnotationFilterFunc(
-			networking.IngressClassAnnotationKey, config.KourierIngressClassName, false,
-		),
-		Handler: controller.HandleAll(impl.Enqueue),
-	}
-
-	ingressInformer.Informer().AddEventHandler(ingressInformerHandler)
-
-	endpointsInformer.Informer().AddEventHandler(controller.HandleAll(
-		controller.EnsureTypeMeta(
-			c.tracker.OnChanged,
-			v1.SchemeGroupVersion.WithKind("Endpoints"),
-		),
-	))
-
-	return impl
+	reconciler.EnvoyXDSServer.MarkIngressesReady(
+		[]*v1alpha1.Ingress{ingress},
+		reconciler.CurrentCaches.SnapshotVersion(),
+	)
 }
